header 반응형에서 login 안했을시 생기는 문제 해결, ãmain bseesartist product 생성, 숫자 단위 , 추가

+++ src/main/java/com/example/ArtAuction_24/domain/product/service/AuctionProductService.java
@@ -53,7 +53,7 @@ public class AuctionProductService {
     }
 
     public void create(String title, String description, String medium, String dimensions,
-                       int startingPrice, int currentBid,
+                       BigDecimal startingPrice, BigDecimal currentBid,
                        LocalDateTime auctionStartDate, String thumbnailImg, String category, Artist artist) {
         AuctionProduct p = AuctionProduct.builder()
                 .title(title)
